requirements.txt
asgiref==3.9.1
boto3==1.40.35
botocore==1.40.35
dj-database-url==3.0.1
Django==4.2.24
django-extensions==4.1
django-storages==1.14.6
django-widget-tweaks==1.5.0
gunicorn==23.0.0
jmespath==1.0.1
packaging==25.0
pillow==11.3.0
psycopg==3.2.10
psycopg-binary==3.2.10
python-dateutil==2.9.0.post0
qrcode==8.2
s3transfer==0.14.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
urllib3==2.5.0
